refactor(oa): make optimize_anything substantive; warn instead of raise on no budget
- optimize_anything now builds the Task + server and runs inline rather
  than delegating to optimize_anything_from_task (which is now the thin
  Task-unpacking wrapper).
- A missing budget (neither max_evals nor max_token_cost set) now warns
  and runs unbounded instead of raising ValueError.
- Move optimize_anything_from_task and optimize_anything_with_server into
  oa/ensemble.py (still re-exported from gepa.optimize_anything); they
  delegate to optimize_anything / the run primitives via lazy imports.
